NPYParser quick fix.

+++ strlearn/streams/NPYParser.py
@@ -61,7 +61,7 @@ class NPYParser:
     def _make_classification(self):
         # Read CSV
         ds = np.load(self.path)
-        self.classes_ = np.unique(ds[:,-1])
+        self.classes_ = np.unique(ds[:,-1]).astype(int)
         return ds[:,:-1], ds[:,-1]
 
     def __str__(self):
